+++ tests/testthat/test-ANOPA-2.R
@@ -1,11 +1,11 @@
-context("ANOPA:: Testing main function (3 sections)")
+context("ANOPA:: Testing anopa() function (3 sections)")
     # expect_output( str(res), "data.frame")
     # expect_equal( "ggplot" %in% class(plt), TRUE)
     # expect_message( p <- superbPlot(dta2a, etc ))
     # expect_error(), expect_warning(), expect_condition( , class = "")
 
 
-test_that("Testing anopa on simple data sets (1/3)", {
+test_that("Testing anopa on 5 simple data set designs (1/3)", {
 
 	old <- options() 
 	on.exit(options(old)) 
@@ -39,7 +39,6 @@ test_that("Testing anopa on simple data sets (1/3)", {
 		expect_equal( dim(r1Bb$wideData), c(175,2))
 		expect_equal( round(r1Bb$omnibus[1,1]-0.03238,7), 0, tol=0.00001)
 
-
 	frml1B <- Value ~ state * Variable | Id
 	r1Bc <- anopa(frml1B, l1B)
 		expect_equal( dim(r1Bc$wideData), c(175,2))
@@ -52,6 +51,11 @@ test_that("Testing anopa on simple data sets (1/3)", {
 
 
 	# all the examples with one WS factor (w/l)			DONE
+	frmc1W <- {cbind(bpre, b1week, b5week); n; uAlpha} ~ .
+	r1Wa <- anopa(frmc1W, c1W, WSFactors = "moment(3)")
+		expect_equal( dim(r1Wa$wideData), c(19,3))
+		expect_equal( round(r1Wa$omnibus[1,1]-0.003512184,7), 0, tol=0.00001)
+
 	frmw1W <- cbind(bpre, b1week, b5week) ~ .
 	r1Wb <- anopa(frmw1W, w1W, WSFactors = "moment(3)")
 		expect_equal( dim(r1Wb$wideData), c(19,3))
@@ -84,6 +88,12 @@ test_that("Testing anopa on simple data sets (1/3)", {
 
 
 	# all the examples with two WS factors (w/l)		DONE
+	frmc2W <- {cbind(r11,r12,r13,r21,r22,r23);n;uAlpha} ~ .
+	r2Wa <- anopa(frmc2W, c2W, WSFactors = c("G(3)", "F(2)"))
+		expect_equal( dim(r2Wa$wideData), c(30,6))
+		expect_equal( round(r2Wa$omnibus[1,1]-0.0502356,7), 0, tol=0.00001)
+		expect_equal( round(r2Wa$omnibus[2,1]-0.00732517,7), 0, tol=0.00001)
+
 	frmw2W <- cbind(r11,r12,r13,r21,r22,r23) ~ .
 	r2Wb <- anopa(frmw2W, w2W, WSFactors = c("G(3)", "F(2)"))
 		expect_equal( dim(r2Wb$wideData), c(30,6))
@@ -97,7 +107,13 @@ test_that("Testing anopa on simple data sets (1/3)", {
 		expect_equal( round(r2Wc$omnibus[2,1]-0.00732517,7), 0, tol=0.00001)
 
 
-	# all the examples with BW factors (w/l)		HERE
+	# all the examples with BW factors (w/l)		    DONE
+	frmcWB <- {cbind(bpre, bpost, b1week, b5week); n; uAlpha} ~ Status
+	rWBa <- anopa(frmcWB, cWB, WSFactors = "Moment(4)" )
+		expect_equal( dim(rWBa$wideData), c(27,5))
+		expect_equal( round(rWBa$omnibus[1,1]-0.006472,7), 0, tol=0.00001)
+		expect_equal( round(rWBa$omnibus[2,1]-0.174546,7), 0, tol=0.00001)
+        
 	frmwWB <- cbind(bpre, bpost, b1week, b5week) ~ Status
 	rWBb <- anopa(frmwWB, wWB, WSFactors = "Moment(4)" )
 		expect_equal( dim(rWBb$wideData), c(27,5))
@@ -183,5 +199,34 @@ test_that("Testing simulated data sets (3/3)", {
 	pwr <- mean(dec)
 	expect_equal( pwr, .80, tolerance = 0.05)
 
+
+})
+
+
+test_that("Testing two formats of the same data (3/3)", {
+
+	old <- options() 
+	on.exit(options(old)) 
+	options(ANOPA.feedback = 'none')
+
+    df <- GRP( props = c(.5, .5, .5, .5, .5, .7, .7, .2), n = c(81,21,16,30), 
+        BSDesign = list( group = c("control","treatment1","treatment2","treament3") ),
+        WSDesign = list( foraging = c("before", "after") ),
+        sname = "frg"
+    )
+
+    ## OLD: anopa on the wide format
+    w <- anopa( cbind( frg.before, frg.after) ~ group, df, WSFactors= "frg(2)")
+    dfC <- toCompiled(w)
+    print(pltw <- anopaPlot(w))
+    expect_equal( "ggplot" %in% class(pltw), TRUE)
+
+    ## NEW: compiled format with repeated measures
+    v <- anopa( {cbind(frg.before,frg.after); Count; uAlpha} ~ group, dfC, WSFactors = "frg(2)")
+    print(pltv <- anopaPlot(v))
+    expect_equal( "ggplot" %in% class(pltv), TRUE)
+
+    ## both returns the same response
+    expect_equal( max(abs(summary(w) - summary(v)), na.rm = TRUE), 0, tol = 0.000001)
 
 })